Initial commit with a wireframe
Upon finish analyse the assignment brief and the API specifications, a wirefram is made in a pretotype template (by using Pencil Evolus).
The index.html is a webpage contains clickables to showcase the app.
In current design, There are 9 pages to showcase features that are going to be in the app. Including Login, Logout, Sign up, create chit, Follower management, Update user details, Check users' details and load chit list. They will be more features adding into the app in the future, so the pretotype is suspect to change.