Implemented basic navigation structure
Added navigators.js to implement the main navigation structure. Screens are stored in a folder called 'screens'. The navigators are consist of three stack navigators, one tab navigator and a draw navigator. The draw navigator are nested with two stack navigators and one tab navigaor while the tab navigator  are nested with a stack navigator. The structure might change in the future.
The wireframe will be changed, due to format of presentation is wrong and navigation structure not included.
Will work on GET/Chits endpoint to fill HomeScreen with Chits in the future.