Implemented GET/search_user endpoint
Changed the class name in PostChit.js.
Altered Search.js to implement GET/search_user, everytime user type in the search bar a request will be send to retrieve the users that match the keyword and every list item is clickable, will add a function redirect to that user's profile.
There is a bug when user hit the backspace too fast to clear up the search bar, the result remain on the list. The only exception is when user hit the backspace slowly or click the built-in clear button (an 'x' at the right side). There is another bug where image are not loaded in the avatar bubble, which will fix it when implementing image management.
Next will start the next part of the follower management, 'follow a user'.