Fully implemented other follower management endpoints
Added two new screens in navigator.js to render followers and following list.
Both screens can be access from OtherProfile.js by clicking the word 'Followers' or 'Following'.
Upon clicking the users in the 'Followers'/'Following' list, will direct user to the specific user profile. If user tried to reach his/her profile through this way will see that the edit button and logout button is gone, only follow button will appear on the header. Will fix this bug in the polishing phase, because this is not a major problem that cause app crash. Other problems such as image not showing in avatar will be fix on the next step in the progress, numbers beside 'Followers' and 'Following' also need a fix because it supposed to be updating the count when user click the 'Follow' button.