Impleented scrollable chit list in profiles' screen and home screen
Added scrollable lists in profile.js, otherprofile.js and homescreen.js. The timestamp also has been converted to readable date format instead of a long string of milliseconds. There are some problems need to be fix, the Flatlist needs flex: 1 to able to scroll when other components are in the view and the lists' position need to amend in the future to be more dynamic. Currently the chits in the list are clickable but only return messages need to add a screen to show chit details.
More styling work to do since the profiles' styling is hard-coded.
Will add image management in the next commit.