Implemented image management for user profile picture
Added get photo and post photo endpoints for profile.js and editprofile.js, editphoto.js is a camera to capture and upload to server. Will add an upload button in the future to let user preview first then upload instead of just send it immediately.
There is a warning shows which is source.uri cannot be empty that needs to fix.
Added a new component that provide a base URL so that IP address can be change globally.