Added upload image and getting location in postchit.js
Implemented get geolocation in postchit.js only a string of coords shows on screen now. Need react-native-maps to picture it then post it to server.
Image upload for chits is complete, but still deciding how to retrieve it and show it in homepage.

Error report:
After installing react-native-geolocation-service, app can't be install in emulator, the error message is: Execution failed for task ':app:processDebugResources'
The solution is go cmd and run these:
1. cd android
2. gradlew.bat clean
Source: https://github.com/oblador/react-native-vector-icons/issues/429

There is also a problem to show picture save through react-native-image-picker, although the base64 string is similar but it doesn't render as expected. Will change back to RNCamera instead if no solution is found.