Implemented image and location posting in postChit.js and fixed some bugs
Added image posting in postChit.js with react-native-camera instead of using react-native-image-picker because the way it format the data causes data to corrupt when user tries to retrieve.
Enabled multidex because the app is over '64K reference limit'.
Added another screen, 'capturePhoto' for capturing photo and send a preview to PostChit.js.
Amended how state stores the location/coordinates and user will be sending longitude and latitude to the server.
Will be showing photo in chits in the homepage and showing places instead of longitude and latitude.
Bug(s) fixed:
Counter didn't reset after posting.
Image was not reset after posting.