Implemented reverse geocoding for location feature and improved UI in postChit.js
Added react-native-geocoding for reverse geocoding to get human-readable location instead of longitude and latitude.
Added small 'x' icon beside photo preview and location to reset state to null, so user can decide to send chits with or without location and photo.
Refactored code to show certain component when state is not null.
Uninstalled react-native-image-picker because not going to use it.
Will work on homeScreen to show location and image next.