Implemented show image and pull to refresh in homescreen.js
Added a for loop to loop through chit list and extract photo/image.
Wrapped the file reader with a promise wrapper to make it asynchronous.
Added pull to refresh feature in home screen to refresh page because componentDidMount only triggered once.
Side not:e: calling setState in render() will cause infinite loop
Will update the cards to show who post, when does it post and tagged location (if there's any).