Styled the home screen
Added geocoder to show location in human-readable format.
Added components in <Card> to show location, timestamp and users' name.
Added some styling for the components too.
Styled the <Text> for name to wrap around when it is too long, so the time stamp wouldn't go off screen.
Implemented location extraction in the for loop where user get the photos.