Implemented draft management and logo on login page
Removed and added an image in /images folder.
Added default avatar picture in home screen, because getting profile picture is still under development.
Added a logo in login page.
Implemented draft manaement with AsyncStorage, for now the only way to save and delete draft is to tap on the red cross on the top right corner and only text of the post is saved in the draft. Expected to show delete or save options when user try to reach other page other than photo capturing.
Hope to add placeholder like '(no context)' for draft with no text.
Remove some header of stack navigators.
Bug(s):
Need to handle error if user is not logged in and try to save draft.
Several error is not handle carefully.